Define Model Info in a separate class
